refactor: deprecate header-only style
use CMake object library
