+++ CMakeLists.txt
@@ -15,6 +15,8 @@ include_directories(${PROJECT_SOURCE_DIR}/include)
 
 find_package(mimalloc 2.0 REQUIRED)
 
+add_subdirectory(./lib)
+
 add_subdirectory(./test)
 add_subdirectory(./example)
 # add_subdirectory(./src)
