+++ example/fingerprint_blocking.cpp
@@ -1,8 +1,8 @@
-#include <mimalloc-new-delete.h>
+// #include <mimalloc-new-delete.h>
 #include "co_context.hpp"
 #include "co_context/net/acceptor.hpp"
 #include "co_context/buffer.hpp"
-
+#include "co_context/lazy_io.hpp"
 #include <filesystem>
 #include <fcntl.h>
