+++ example/netcat.cpp
@@ -1,4 +1,4 @@
-#include <mimalloc-new-delete.h>
+// #include <mimalloc-new-delete.h>
 #include "co_context.hpp"
 #include <thread>
