+++ example/nop.cpp
@@ -1,4 +1,3 @@
-#include <mimalloc-new-delete.h>
 #include "co_context.hpp"
 #include "co_context/net/acceptor.hpp"
 #include "co_context/buffer.hpp"
