+++ example/rand4kw_callback.cpp
@@ -1,14 +1,13 @@
-#include <mimalloc-new-delete.h>
+// #include <mimalloc-new-delete.h>
 #include "co_context.hpp"
 #include "co_context/net/acceptor.hpp"
 #include "co_context/buffer.hpp"
-
+#include "co_context/lazy_io.hpp"
 #include <filesystem>
 #include <random>
 #include <fcntl.h>
 #include <atomic>
 #include <chrono>
-#include <semaphore>
 
 using namespace co_context;
