+++ include/co_context.hpp
@@ -19,25 +19,13 @@
  */
 #pragma once
 
-#include <new>
-#include <memory>
-#include <atomic>
-#include <semaphore>
 #include <thread>
-#include <chrono>
-#include <cassert>
 #include <vector>
 #include <queue>
 #include "uring.hpp"
-#include "co_context/task.hpp"
 #include "co_context/config.hpp"
-#include "co_context/set_cpu_affinity.hpp"
 #include "co_context/task_info.hpp"
 #include "co_context/main_task.hpp"
-#include "co_context/log/log.hpp"
-
-#include <iostream>
-#include <syncstream>
 
 namespace co_context {
 
@@ -66,7 +54,7 @@ namespace detail {
         uint32_t tid;
     };
 
-    inline static thread_local thread_meta this_thread;
+    extern thread_local thread_meta this_thread;
 
     struct alignas(cache_line_size) worker_meta {
         enum class worker_state : uint8_t { running, idle, blocked };
@@ -94,7 +82,7 @@ namespace detail {
         swap_cur reap_cur;
         std::queue<task_info *> submit_overflow_buf;
 
-        void submit(task_info *io_info) noexcept;
+        void submit(task_info_ptr io_info) noexcept;
 
         std::coroutine_handle<> schedule() noexcept;
 
@@ -102,17 +90,7 @@ namespace detail {
 
         void co_spawn(main_task entrance) noexcept;
 
-        void run(const int thread_index, io_context *const context) {
-            init(thread_index, context);
-            log::v("worker[%d] run...\n", thread_index);
-
-            while (true) {
-                auto coro = this->schedule();
-                log::v("worker[%d] found coro to run\n", thread_index);
-                coro.resume();
-                log::v("worker[%d] idle\n", thread_index);
-            }
-        }
+        void run(const int thread_index, io_context *const context);
     };
 
 } // namespace detail
@@ -124,19 +102,10 @@ inline constexpr uint16_t task_info_number_per_cache_line =
 
 class [[nodiscard]] io_context final {
   public:
-    /*
-    static constexpr uint16_t swap_slots =
-        config::swap_capacity / task_info_number_per_cache_line;
-
-    static constexpr uint16_t task_info_number_per_thread =
-        task_info_number_per_cache_line * swap_slots;
-    */
-
     /**
      * One thread for io_context submit/reap. Another thread for io_uring (if
      * needed).
      */
-
     static_assert(config::worker_threads_number >= 1);
 
     using uring = liburingcxx::URing<config::io_uring_flags>;
@@ -158,67 +127,13 @@ class [[nodiscard]] io_context final {
         uint16_t tid = 0;
         uint16_t off = 0;
 
-        inline void next() {
-            if (++tid == config::worker_threads_number) [[unlikely]] {
-                tid = 0;
-                off = (off + 1) % config::swap_capacity;
-            }
-        }
-
-        inline bool try_find_empty(const swap_zone &swap) noexcept {
-            constexpr uint32_t swap_size =
-                sizeof(swap_zone) / sizeof(task_info_ptr);
-            int i = 0;
-            while (i < swap_size && swap[tid][off] != nullptr) {
-                ++i;
-                next();
-            }
-            if (i != swap_size) [[likely]] {
-                return true;
-            } else {
-                next();
-                return false;
-            }
-        }
-
-        // inline void wait_for_empty(const swap_zone &swap) noexcept {
-        //     while (swap[tid][off] != nullptr) next();
-        // }
-
-        inline bool try_find_exist(const swap_zone &swap) noexcept {
-            constexpr uint32_t swap_size =
-                sizeof(swap_zone) / sizeof(task_info_ptr);
-            int i = 0;
-            while (i < swap_size && swap[tid][off] == nullptr) {
-                ++i;
-                next();
-            }
-            if (i != swap_size) [[likely]] {
-                return true;
-            } else {
-                next();
-                return false;
-            }
-        }
-
-        // inline void wait_for_exist(const swap_zone &swap) noexcept {
-        //     while (swap[tid][off] == nullptr) next();
-        // }
-
+        void next() noexcept;
+        bool try_find_empty(const swap_zone &swap) noexcept;
+        bool try_find_exist(const swap_zone &swap) noexcept;
         // release the task into the swap zone
-        inline void
-        release(swap_zone &swap, task_info_ptr task) const noexcept {
-            std::atomic_store_explicit(
-                reinterpret_cast<std::atomic<task_info_ptr> *>(&swap[tid][off]),
-                task, std::memory_order_release);
-        }
-
-        inline void
-        release_relaxed(swap_zone &swap, task_info_ptr task) const noexcept {
-            std::atomic_store_explicit(
-                reinterpret_cast<std::atomic<task_info_ptr> *>(&swap[tid][off]),
-                task, std::memory_order_relaxed);
-        }
+        void release(swap_zone &swap, task_info_ptr task) const noexcept;
+        void
+        release_relaxed(swap_zone &swap, task_info_ptr task) const noexcept;
     };
 
     // place main thread's high frequency data here
@@ -228,8 +143,6 @@ class [[nodiscard]] io_context final {
     std::queue<task_info_ptr> reap_overflow_buf;
 
     alignas(cache_line_size) char __cacheline_barrier[64];
-
-    // std::counting_semaphore<1> idle_worker_quota{1};
 
   public:
     using worker_meta = detail::worker_meta;
@@ -243,129 +156,27 @@ class [[nodiscard]] io_context final {
     const unsigned ring_entries;
 
   private:
-    void forward_task(task_info_ptr task) noexcept {
-        // TODO optimize scheduling strategy
-        if (r_cur.try_find_empty(reap_swap)) [[likely]] {
-            reap_swap[r_cur.tid][r_cur.off] = task;
-            r_cur.next();
-            log::v("ctx forward_task to [%u][%u]\n", r_cur.tid, r_cur.off);
-        } else {
-            reap_overflow_buf.push(task);
-            log::d("ctx forward_task to reap_OF\n");
-        }
-    }
+    void forward_task(task_info_ptr task) noexcept;
 
-    bool try_submit(task_info_ptr task) noexcept {
-        if (task->type == task_info::task_type::co_spawn) {
-            forward_task(task);
-            return true;
-        }
-
-        if (ring.SQSpaceLeft() == 0) [[unlikely]] {
-            log::d("ctx try_submit failed SQfull\n");
-            return false; // SQRing is full
-        }
-
-        liburingcxx::SQEntry *sqe = ring.getSQEntry();
-
-        sqe->cloneFrom(*task->sqe);
-        ring.submit();
-
-        log::v("ctx submit to ring\n");
-
-        return true;
-    }
+    bool try_submit(task_info_ptr task) noexcept;
 
     /**
      * @brief poll the submission swap zone
      * @return if submit_swap capacity might be healthy
      */
-    bool poll_submission() noexcept {
-        // submit round
-        if (!s_cur.try_find_exist(submit_swap)) { return false; }
+    bool poll_submission() noexcept;
 
-        task_info_ptr const io_info = submit_swap[s_cur.tid][s_cur.off];
-        submit_swap[s_cur.tid][s_cur.off] = nullptr;
-        log::v("ctx poll_submission at [%u][%u]\n", s_cur.tid, s_cur.off);
-        s_cur.next();
+    bool try_clear_submit_overflow_buf() noexcept;
 
-        if (try_submit(io_info)) [[likely]] {
-            return true;
-        } else {
-            submit_overflow_buf.push(io_info);
-            log::d("ctx poll_submission failed submit_OF\n");
-            return false;
-        }
-    }
-
-    inline bool try_clear_submit_overflow_buf() noexcept {
-        if (submit_overflow_buf.empty()) return true;
-        do {
-            task_info_ptr task = submit_overflow_buf.front();
-            // OPTIMIZE impossible for task_type::co_spawn
-            if (try_submit(task)) {
-                submit_overflow_buf.pop();
-            } else {
-                log::d("ctx try_clear_submit (partially) failed\n");
-                return false;
-            }
-        } while (!submit_overflow_buf.empty());
-        log::d("ctx try_clear_submit succ\n");
-        return true;
-    }
-
-    bool try_reap(task_info_ptr task) noexcept {
-        if (!r_cur.try_find_empty(reap_swap)) [[unlikely]] {
-            log::d("ctx try_reap failed reap_swap is full\n");
-            return false;
-        }
-
-        r_cur.release(reap_swap, task);
-        log::v("ctx try_reap at [%u][%u]\n", r_cur.tid, r_cur.off);
-        r_cur.next();
-        return true;
-    }
+    bool try_reap(task_info_ptr task) noexcept;
 
     /**
      * @brief poll the completion swap zone
      * @return if load exists and capacity of reap_swap might be healthy
      */
-    bool poll_completion() noexcept {
-        // reap round
-        liburingcxx::CQEntry *polling_cqe = ring.peekCQEntry();
-        if (polling_cqe == nullptr) return false;
+    bool poll_completion() noexcept;
 
-        log::v("ctx poll_completion found\n");
-
-        task_info_ptr io_info =
-            reinterpret_cast<task_info_ptr>(polling_cqe->getData());
-        io_info->result = polling_cqe->getRes();
-        ring.SeenCQEntry(polling_cqe);
-
-        if (try_reap(io_info)) [[likely]] {
-            return true;
-        } else {
-            reap_overflow_buf.push(io_info);
-            log::d("ctx poll_completion failed reap_OF\n");
-            return false;
-        }
-    }
-
-    inline bool try_clear_reap_overflow_buf() noexcept {
-        if (reap_overflow_buf.empty()) return true;
-        do {
-            task_info_ptr task = reap_overflow_buf.front();
-            // OPTIMIZE impossible for task_type::co_spawn
-            if (try_reap(task)) {
-                reap_overflow_buf.pop();
-            } else {
-                log::d("ctx try_clear_reap (partially) failed\n");
-                return false;
-            }
-        } while (!reap_overflow_buf.empty());
-        log::d("ctx try_clear_reap succ\n");
-        return true;
-    }
+    bool try_clear_reap_overflow_buf() noexcept;
 
     [[noreturn]] void stop() noexcept {
         log::i("ctx stopped\n");
@@ -388,83 +199,17 @@ class [[nodiscard]] io_context final {
     io_context(unsigned io_uring_entries, uring::Params &&io_uring_params)
         : io_context(io_uring_entries, io_uring_params) {}
 
-    void init() noexcept {
-        memset(submit_swap, 0, sizeof(submit_swap));
-        memset(reap_swap, 0, sizeof(reap_swap));
-        detail::this_thread.ctx = this;
-        detail::this_thread.tid = std::thread::hardware_concurrency() - 1;
-    }
+    void init() noexcept;
 
-    void probe() const {
-        using namespace std;
-        using namespace co_context::config;
-        cout << "number of logic cores: " << std::thread::hardware_concurrency()
-             << endl;
-        cout << "size of io_context: " << sizeof(io_context) << endl;
-        cout << "size of uring: " << sizeof(uring) << endl;
-        cout << "size of two swap_zone: " << 2 * sizeof(swap_zone) << endl;
-        // cout << "swap_zone is_lock_free: "
-        //      << submit_swap[0][0][0].is_lock_free() << endl;
-        cout << "number of worker_threads: " << worker_threads_number << endl;
-        cout << "swap_capacity per thread: " << swap_capacity << endl;
-        cout << "size of single worker_meta: " << sizeof(worker_meta) << endl;
-    }
+    void probe() const;
 
-    void make_thread_pool() {
-        for (int i = 0; i < config::worker_threads_number; ++i)
-            worker[i].sharing.host_thread =
-                std::thread{&worker_meta::run, worker + i, i, this};
-    }
+    void make_thread_pool();
 
-    void co_spawn(main_task entrance) {
-        const uint32_t tid = detail::this_thread.tid;
-        if (tid < config::worker_threads_number)
-            return worker[tid].co_spawn(entrance);
-        if (r_cur.try_find_empty(reap_swap)) [[likely]] {
-            r_cur.release(reap_swap, entrance.get_io_info_ptr());
-            log::d("ctx co_spawn at [%u][%u]\n", r_cur.tid, r_cur.off);
-            r_cur.next();
-        } else {
-            reap_overflow_buf.push(entrance.get_io_info_ptr());
-            log::d("ctx co_spawn failed reap_OF\n");
-        }
-    }
+    void co_spawn(main_task entrance);
 
     void can_stop() noexcept { will_stop = true; }
 
-    [[noreturn]] void run() {
-        detail::this_thread.worker = nullptr;
-#ifdef USE_CPU_AFFINITY
-        detail::this_thread.tid = std::thread::hardware_concurrency() - 1;
-        detail::set_cpu_affinity(detail::this_thread.tid);
-#endif
-        log::v("ctx making thread pool\n");
-        make_thread_pool();
-        log::v("ctx make_thread_pool end\n");
-
-        while (!will_stop) [[likely]] {
-                log::v("ctx polling\n");
-                if (try_clear_submit_overflow_buf()) {
-                    for (uint8_t i = 0; i < config::submit_poll_rounds; ++i) {
-                        // printf("submit start!\n");
-                        if (!poll_submission()) break;
-                        // printf("submitted!\n");
-                        // poll_submission();
-                    }
-                }
-
-                if (try_clear_reap_overflow_buf())
-                    for (uint8_t i = 0; i < config::reap_poll_rounds; ++i) {
-                        if (!poll_completion()) break;
-                        // printf("reapped!\n");
-                        // poll_completion();
-                    }
-
-                // std::this_thread::sleep_for(std::chrono::milliseconds(100));
-            }
-
-        this->stop();
-    }
+    [[noreturn]] void run();
 
     inline uring &io() noexcept { return ring; }
 
@@ -495,130 +240,5 @@ inline void co_context_stop() noexcept {
 inline uint32_t co_get_tid() noexcept {
     return detail::this_thread.tid;
 }
-
-namespace detail {
-
-    inline void worker_meta::swap_cur::next() noexcept {
-        off = (off + 1) % co_context::config::swap_capacity;
-    }
-
-    inline bool
-    worker_meta::swap_cur::try_find_empty(swap_zone &swap) noexcept {
-        uint16_t i = 0;
-        const uint32_t tid = this_thread.tid;
-        while (i < config::swap_capacity && swap[tid][off] != nullptr) {
-            next();
-        }
-        if (i == config::swap_capacity) [[unlikely]] {
-            next();
-            return false;
-        }
-        return true;
-    }
-
-    inline bool
-    worker_meta::swap_cur::try_find_exist(swap_zone &swap) noexcept {
-        uint16_t i = 0;
-        const uint32_t tid = this_thread.tid;
-        while (i < config::swap_capacity && swap[tid][off] == nullptr) {
-            next();
-        }
-        if (i == config::swap_capacity) [[unlikely]] {
-            next();
-            return false;
-        }
-        return true;
-    }
-
-    inline void worker_meta::swap_cur::release(
-        swap_zone &swap, task_info_ptr task) const noexcept {
-        const uint32_t tid = this_thread.tid;
-        std::atomic_store_explicit(
-            reinterpret_cast<std::atomic<task_info_ptr> *>(&swap[tid][off]),
-            task, std::memory_order_release);
-    }
-
-    inline void worker_meta::swap_cur::release_relaxed(
-        swap_zone &swap, task_info_ptr task) const noexcept {
-        const uint32_t tid = this_thread.tid;
-        std::atomic_store_explicit(
-            reinterpret_cast<std::atomic<task_info_ptr> *>(&swap[tid][off]),
-            task, std::memory_order_relaxed);
-    }
-
-    inline void
-    worker_meta::init(const int thread_index, io_context *const context) {
-        assert(submit_overflow_buf.empty());
-        detail::this_thread.ctx = context;
-        detail::this_thread.worker = this;
-        detail::this_thread.tid = thread_index;
-#ifdef USE_CPU_AFFINITY
-        const unsigned logic_cores = std::thread::hardware_concurrency();
-        if constexpr (config::using_hyper_threading) {
-            if (thread_index * 2 < logic_cores) {
-                detail::set_cpu_affinity(thread_index * 2);
-            } else {
-                detail::set_cpu_affinity(thread_index * 2 % logic_cores + 1);
-            }
-        } else {
-            detail::set_cpu_affinity(thread_index);
-        }
-#endif
-    }
-
-    inline void worker_meta::co_spawn(main_task entrance) noexcept {
-        this->submit(entrance.get_io_info_ptr());
-    }
-
-    inline void worker_meta::submit(task_info_ptr io_info) noexcept {
-        const unsigned tid = this_thread.tid;
-        auto &ctx = *this_thread.ctx;
-        auto &cur = this->submit_cur;
-        // std::cerr << cur.slot << " " << tid << cur.off << std::endl;
-        if (cur.try_find_empty(ctx.submit_swap)) [[likely]] {
-            cur.release(ctx.submit_swap, io_info);
-            log::v("worker[%u] submit at [%u]\n", tid, cur.off);
-            cur.next();
-        } else {
-            this->submit_overflow_buf.push(io_info);
-            log::d("worker[%u] submit to OF\n", tid);
-        }
-    }
-
-    inline std::coroutine_handle<> worker_meta::schedule() noexcept {
-        const uint32_t tid = this_thread.tid;
-        auto &ctx = *this_thread.ctx;
-        auto &r_cur = this->reap_cur;
-
-        log::v("worker[%u] scheduling\n", tid);
-
-        while (true) {
-            // handle overflowed submission
-            if (!this->submit_overflow_buf.empty()) {
-                auto &s_cur = this->submit_cur;
-                if (s_cur.try_find_empty(ctx.submit_swap)) {
-                    task_info_ptr task = this->submit_overflow_buf.front();
-                    this->submit_overflow_buf.pop();
-                    s_cur.release_relaxed(ctx.submit_swap, task);
-                    log::d(
-                        "worker[%u] submit from OF_buf to [%u]\n", tid,
-                        s_cur.off);
-                    s_cur.next();
-                }
-            }
-
-            if (r_cur.try_find_exist(ctx.reap_swap)) {
-                const task_info_ptr io_info = ctx.reap_swap[tid][r_cur.off];
-                ctx.reap_swap[tid][r_cur.off] = nullptr;
-                log::d("worker[%u] found [%u]\n", tid, r_cur.off);
-                r_cur.next();
-                // printf("get task!\n");
-                return io_info->handle;
-            }
-            // TODO consider tid_hint here
-        }
-    }
-
-} // namespace detail
 
 } // namespace co_context
