+++ include/co_context/buffer.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <concepts>
+// #include <concepts>
 #include <type_traits>
 #include <span>
