+++ include/co_context/task_info.hpp
@@ -1,7 +1,13 @@
 #pragma once
-#include "../uring.hpp"
 #include <coroutine>
+#include <atomic>
 #include "co_context/log/log.hpp"
+
+
+namespace liburingcxx {
+class SQEntry;
+class CQEntry;
+}
 
 namespace co_context {
 namespace detail {
@@ -14,6 +20,7 @@ namespace detail {
             SQEntry *sqe;
             CQEntry *cqe;
             int32_t result;
+            std::atomic_int_fast32_t *remaining_count;
         };
         std::coroutine_handle<> handle;
         int tid_hint;
@@ -23,28 +30,9 @@ namespace detail {
             log::v("task_info generated\n");
         }
 
-        static task_info nop() noexcept {
-            task_info ret{task_type::nop};
-            ret.sqe = nullptr;
-            ret.handle = nullptr;
-            ret.tid_hint = -1;
-            return ret;
-        }
+        // static task_info nop() noexcept;
 
-        static task_info *new_sqe() {
-            task_info *addr = reinterpret_cast<task_info *>(
-                mi_malloc(sizeof(task_info) + sizeof(SQEntry)));
-            addr->sqe =
-                reinterpret_cast<SQEntry *>((char *)addr + sizeof(task_info));
-            addr->type = task_type::sqe;
-            return addr;
-        }
-
-        /* static task_info from_worker(std::coroutine_handle<> handle)
-           noexcept { return task_info{ .sqe = nullptr, .handle{handle},
-                        .tid_hint = this_thread.tid,
-                    };
-                } */
+        // static task_info *new_sqe();
     };
 
     using task_info_ptr = task_info *;
